feat(db): add entity-of-record columns to subscriptions (ARCH-M1-11 / T-112)
Adds nullable `legal_tenant_id`, `billing_entity`, `governing_law_snapshot`,
and `terms_version_accepted` to the `subscriptions` table via additive
Drizzle migration 0077.  `subscriptions` is confirmed as the M1
system-of-record (no new memberships aggregate).  Existing rows are
backfilled: `legal_tenant_id` from `tenant_id`, `governing_law_snapshot`
from `tenants.governing_law`.  `billing_entity` and `terms_version_accepted`
are null for historical rows and will be populated by the live-writer
capture path in a follow-up slice.

- Extract `subscriptions` table to `schema/subscriptions.ts` (T-112 EOR
  columns + ISO check constraint on `governing_law_snapshot`)
- Extract `membershipPlans` table to `schema/membership-plans.ts` (breaks
  circular dep for subscriptions/membershipPlans)
- `memberships.ts` re-exports both for backward compatibility; all existing
  imports and barrel exports work unchanged
- Migration 0077 adds columns, backfills, FK for `legal_tenant_id`, and
  check constraint for `governing_law_snapshot`
- 3 focused unit tests: schema export, column ordering, migration content
- Bump `maxTrackedFiles` to 4009 (+4 new tracked files)
- Promotes ARCH-M1-11 in current-program, current-tracker, and
  architecture-finalization-tracker; T-112 status WIP

Phase C: no proxy.ts, no route, no auth, no tenancy refactor, no billing
logic change, no Stripe, no README/AGENTS.
