+++ packages/database/src/schema/memberships.ts
@@ -1,90 +1,13 @@
-import {
-  boolean,
-  decimal,
-  index,
-  integer,
-  jsonb,
-  pgTable,
-  text,
-  timestamp,
-  uniqueIndex,
-} from 'drizzle-orm/pg-core';
+import { boolean, index, pgTable, text, timestamp, uniqueIndex } from 'drizzle-orm/pg-core';
 
 import { user } from './auth';
-import { membershipTierEnum, subscriptionStatusEnum } from './enums';
 import { branches } from './rbac';
 import { tenants } from './tenants';
 
-export const membershipPlans = pgTable('membership_plans', {
-  id: text('id').primaryKey(),
-  tenantId: text('tenant_id')
-    .notNull()
-    .references(() => tenants.id),
-  name: text('name').notNull(),
-  description: text('description'),
-  tier: membershipTierEnum('tier').notNull(),
-  interval: text('interval').default('year').notNull(),
-  price: decimal('price', { precision: 8, scale: 2 }).notNull(),
-  currency: text('currency').default('EUR').notNull(),
-  membersIncluded: integer('members_included').default(1).notNull(),
-  legalConsultationsPerYear: integer('legal_consultations_per_year'),
-  mediationDiscountPercent: integer('mediation_discount_percent').default(0).notNull(),
-  successFeePercent: integer('success_fee_percent').default(15).notNull(),
-  paddlePriceId: text('paddle_price_id'),
-  paddleProductId: text('paddle_product_id'),
-  features: jsonb('features').$type<string[]>().notNull().default([]),
-  isActive: boolean('is_active').default(true).notNull(),
-  createdAt: timestamp('created_at').defaultNow().notNull(),
-  updatedAt: timestamp('updated_at')
-    .$onUpdate(() => new Date())
-    .notNull(),
-});
-
-export const subscriptions = pgTable(
-  'subscriptions',
-  {
-    id: text('id').primaryKey(),
-    tenantId: text('tenant_id')
-      .notNull()
-      .references(() => tenants.id),
-    userId: text('user_id')
-      .notNull()
-      .references(() => user.id),
-    status: subscriptionStatusEnum('status').notNull().default('active'),
-    planId: text('plan_id').notNull(),
-    planKey: text('plan_key').references(() => membershipPlans.id),
-    provider: text('provider').default('paddle'),
-    providerSubscriptionId: text('provider_subscription_id'),
-    providerCustomerId: text('provider_customer_id'),
-    currentPeriodStart: timestamp('current_period_start'),
-    currentPeriodEnd: timestamp('current_period_end'),
-    cancelAtPeriodEnd: boolean('cancel_at_period_end').default(false),
-    canceledAt: timestamp('canceled_at'),
-    pastDueAt: timestamp('past_due_at'),
-    gracePeriodEndsAt: timestamp('grace_period_ends_at'),
-    dunningAttemptCount: integer('dunning_attempt_count').default(0),
-    lastDunningAt: timestamp('last_dunning_at'),
-    referredByAgentId: text('referred_by_agent_id').references(() => user.id),
-    referredByMemberId: text('referred_by_member_id').references(() => user.id),
-    referralCode: text('referral_code').unique(),
-    acquisitionSource: text('acquisition_source'),
-    utmSource: text('utm_source'),
-    utmMedium: text('utm_medium'),
-    utmCampaign: text('utm_campaign'),
-    utmContent: text('utm_content'),
-    branchId: text('branch_id').references(() => branches.id), // Branch scoping
-    agentId: text('agent_id').references(() => user.id), // Assigned agent (replaces referredByAgentId)
-    createdAt: timestamp('created_at').defaultNow(),
-    updatedAt: timestamp('updated_at').$onUpdate(() => new Date()),
-  },
-  table => [
-    index('idx_memberships_branch').on(table.branchId),
-    index('idx_memberships_agent').on(table.agentId),
-    uniqueIndex('idx_subscriptions_user').on(table.userId),
-    uniqueIndex('idx_subscriptions_provider_subscription').on(table.providerSubscriptionId),
-    uniqueIndex('subscriptions_tenant_id_id_uq').on(table.tenantId, table.id),
-  ]
-);
+// Extracted to own modules; re-exported here for backward compatibility.
+export { membershipPlans } from './membership-plans';
+import { subscriptions } from './subscriptions';
+export { subscriptions };
 
 export const membershipFamilyMembers = pgTable('membership_family_members', {
   id: text('id').primaryKey(),
@@ -151,7 +74,7 @@ export const membershipCards = pgTable(
       .references(() => user.id),
     subscriptionId: text('subscription_id')
       .notNull()
-      .references(() => subscriptions.id), // Link to specific subscription
+      .references(() => subscriptions.id),
 
     status: text('status', { enum: ['active', 'revoked', 'expired'] })
       .notNull()
@@ -159,7 +82,7 @@ export const membershipCards = pgTable(
 
     // Card Details
     cardNumber: text('card_number').notNull(),
-    qrCodeToken: text('qr_code_token').notNull(), // For generating dynamic QR
+    qrCodeToken: text('qr_code_token').notNull(),
 
     issuedAt: timestamp('issued_at').defaultNow().notNull(),
     expiresAt: timestamp('expires_at'),
